♻️ refactor(embed): restructure type imports
- move embedding related types to a separate module for better organization and modularization
