+++ src/argoproxy/embed.py
@@ -8,7 +8,7 @@ from loguru import logger
 
 from .config import ArgoConfig
 from .constants import EMBED_MODELS
-from .types import CreateEmbeddingResponse, Embedding, Usage
+from .types.embedding import CreateEmbeddingResponse, Embedding, Usage
 from .utils import count_tokens, make_bar, resolve_model_name
 
 DEFAULT_MODEL = "v3small"
