+++ src/argoproxy/types/embedding.py
